requirements.txt
streamlit
boto3
python-dotenv
nest_asyncio
langchain
PyPDF2
langchain-google-genai
faiss-cpu
